Modify and extend the vernac commands.
Revert back to using global references as they are easier to use. They are
also the easiest to convert to dir_paths, and we can require the library using
the located qualid.

Add a command to export all loaded libraries. Currently has a bug when it
tries to export the initial state.

Add a command to set the destination folder (current folder by default).
All the libraries are exported to that folder.